backend/requirements-cloud.txt
fastapi
uvicorn
python-multipart
pydub
opencv-python-headless
ffmpeg-python
scikit-image
numpy
pyspellchecker
google-genai
